refactor: move bar chart form into dedicated component
